refactor(mcp): extract tool definitions to dependency-free module
Move the tool list from index.ts to toolDefinitions.ts so other
packages can import the raw JSON Schema data without pulling in the
MCP SDK, Supabase client, or other runtime dependencies. Enables
cross-package tests that validate the advertised schemas against the
backend's Zod definitions.
