+++ mcp-server/src/index.ts
@@ -1,4 +1,5 @@
 import { Server } from '@modelcontextprotocol/sdk/server/index.js'
+import { tools } from './toolDefinitions.js'
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js'
 import {
 	CallToolRequestSchema,
@@ -11,223 +12,6 @@ import { ApiClient } from './api.js'
 import type { IApiClient } from './api.js'
 import { createToolHandlers, isValidToolName } from './handlers.js'
 import { prompts, getPrompt } from '@matchmaker/shared'
-
-export let tools = [
-	{
-		name: 'add_person',
-		description: 'Add a new person to the matchmaker',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				name: { type: 'string', description: 'Person name' },
-			},
-			required: ['name'],
-		},
-	},
-	{
-		name: 'list_people',
-		description: 'List all people in the matchmaker',
-		inputSchema: {
-			type: 'object',
-			properties: {},
-		},
-	},
-	{
-		name: 'get_person',
-		description: 'Retrieve detailed information about a specific person',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				id: { type: 'string', description: 'Person ID (UUID)' },
-			},
-			required: ['id'],
-		},
-	},
-	{
-		name: 'update_person',
-		description: "Update a person's profile information",
-		inputSchema: {
-			type: 'object',
-			properties: {
-				id: { type: 'string', description: 'Person ID (UUID)' },
-				name: { type: 'string', description: 'Person name' },
-				age: { type: 'number', description: 'Person age' },
-				location: { type: 'string', description: 'Person location' },
-				gender: { type: 'string', description: 'Person gender' },
-				preferences: {
-					type: 'object',
-					properties: {
-						aboutMe: {
-							type: 'object',
-							properties: {
-								height: { type: 'number', description: 'Height in cm' },
-								build: { type: 'string', enum: ['slim', 'average', 'athletic', 'heavy'] },
-								fitnessLevel: { type: 'string', enum: ['active', 'average', 'sedentary'] },
-								ethnicity: { type: 'string' },
-								religion: { type: 'string' },
-								hasChildren: { type: 'boolean' },
-								numberOfChildren: { type: 'number' },
-								isDivorced: { type: 'boolean' },
-								hasTattoos: { type: 'boolean' },
-								hasPiercings: { type: 'boolean' },
-								isSmoker: { type: 'boolean' },
-								occupation: { type: 'string' },
-								income: { type: 'string', enum: ['high', 'moderate', 'low'] },
-							},
-						},
-						lookingFor: {
-							type: 'object',
-							properties: {
-								ageRange: {
-									type: 'object',
-									properties: {
-										min: { type: 'number' },
-										max: { type: 'number' },
-									},
-								},
-								heightRange: {
-									type: 'object',
-									properties: {
-										min: { type: 'number' },
-										max: { type: 'number' },
-									},
-								},
-								fitnessPreference: { type: 'string', enum: ['active', 'average', 'any'] },
-								ethnicityPreference: { type: 'array', items: { type: 'string' } },
-								incomePreference: { type: 'string', enum: ['high', 'moderate', 'any'] },
-								religionRequired: { type: 'string' },
-								wantsChildren: { type: 'boolean' },
-							},
-						},
-						dealBreakers: {
-							type: 'array',
-							items: {
-								type: 'string',
-								enum: ['isDivorced', 'hasChildren', 'hasTattoos', 'hasPiercings', 'isSmoker'],
-							},
-						},
-					},
-				},
-				personality: { type: 'object', description: 'Person personality traits' },
-				notes: { type: 'string', description: 'Notes about the person' },
-			},
-			required: ['id'],
-		},
-	},
-	{
-		name: 'create_introduction',
-		description: 'Create an introduction between two people',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				person_a_id: { type: 'string', description: 'First person ID (UUID)' },
-				person_b_id: { type: 'string', description: 'Second person ID (UUID)' },
-				notes: { type: 'string', description: 'Notes about the introduction' },
-			},
-			required: ['person_a_id', 'person_b_id'],
-		},
-	},
-	{
-		name: 'list_introductions',
-		description: 'List all introductions for the matchmaker',
-		inputSchema: {
-			type: 'object',
-			properties: {},
-		},
-	},
-	{
-		name: 'update_introduction',
-		description: 'Update introduction status or notes',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				id: { type: 'string', description: 'Introduction ID (UUID)' },
-				status: {
-					type: 'string',
-					enum: ['pending', 'accepted', 'declined', 'dating', 'ended'],
-					description: 'Introduction status',
-				},
-				notes: { type: 'string', description: 'Notes about the introduction' },
-			},
-			required: ['id'],
-		},
-	},
-	{
-		name: 'find_matches',
-		description: 'Find compatible matches for a person',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				person_id: { type: 'string', description: 'Person ID (UUID) to find matches for' },
-			},
-			required: ['person_id'],
-		},
-	},
-	{
-		name: 'delete_person',
-		description: 'Soft-delete a person (sets active=false)',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				id: { type: 'string', description: 'Person ID (UUID)' },
-			},
-			required: ['id'],
-		},
-	},
-	{
-		name: 'get_introduction',
-		description: 'Get details of a specific introduction',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				id: { type: 'string', description: 'Introduction ID (UUID)' },
-			},
-			required: ['id'],
-		},
-	},
-	{
-		name: 'submit_feedback',
-		description: 'Submit feedback about an introduction',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				introduction_id: { type: 'string', description: 'Introduction ID (UUID)' },
-				from_person_id: {
-					type: 'string',
-					description: 'Person ID (UUID) submitting the feedback',
-				},
-				content: { type: 'string', description: 'Feedback content' },
-				sentiment: {
-					type: 'string',
-					description: 'Feedback sentiment (e.g., positive, negative, neutral)',
-				},
-			},
-			required: ['introduction_id', 'from_person_id', 'content'],
-		},
-	},
-	{
-		name: 'list_feedback',
-		description: 'Get all feedback for a specific introduction',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				introduction_id: { type: 'string', description: 'Introduction ID (UUID)' },
-			},
-			required: ['introduction_id'],
-		},
-	},
-	{
-		name: 'get_feedback',
-		description: 'Get a specific feedback record',
-		inputSchema: {
-			type: 'object',
-			properties: {
-				id: { type: 'string', description: 'Feedback ID (UUID)' },
-			},
-			required: ['id'],
-		},
-	},
-]
 
 export function createServer(apiClient: IApiClient) {
 	let server = new Server(
